requirements.txt
django==5.2.8
pymysql
gunicorn
cryptography
pillow
requests
urllib3
